+++ src/themachinethatgoesping/tools/vectorinterpolators/akimainterpolator.hpp
@@ -262,7 +262,7 @@ class AkimaInterpolator : public I_Interpolator<double>
                 return;
             }
 
-            for (unsigned int i = 0; i < X.size(); ++i)
+            for (size_t i = 0; i < X.size(); ++i)
             {
                 _akima_spline.push_back(X[i], Y[i]);
 
@@ -313,11 +313,11 @@ class AkimaInterpolator : public I_Interpolator<double>
         XY.reserve(_X.size() + X.size());
 
         // sort _X and _Y by _X (ascending)
-        for (unsigned int i = 0; i < _X.size(); ++i)
+        for (size_t i = 0; i < _X.size(); ++i)
         {
             XY.push_back(std::pair<double, double>(_X[i], _Y[i]));
         }
-        for (unsigned int i = 0; i < X.size(); ++i)
+        for (size_t i = 0; i < X.size(); ++i)
         {
             XY.push_back(std::pair<double, double>(X[i], Y[i]));
         }
@@ -330,7 +330,7 @@ class AkimaInterpolator : public I_Interpolator<double>
         std::vector<double> Y_new;
         X_new.resize(XY.size());
         Y_new.resize(XY.size());
-        for (unsigned int i = 0; i < XY.size(); ++i)
+        for (size_t i = 0; i < XY.size(); ++i)
         {
             X_new[i] = XY[i].first;
             Y_new[i] = XY[i].second;
